requirements/base.txt
amulet-core==1.6
black==22.3.0
pylint==2.12.2
PySide6==6.2.2
QtDesign6==0.0.1
